Qeuing Emails
php artisan queue:work to run the queue worker. Typically run as a separate process from your main Laravel application. It continuously listens for jobs in the queue and processes them. In production, you usually configure Supervisor (on Linux) to run and monitor queue:work so that it restarts automatically if it crashes.