[EXERCISE] PUT /albums/{id}
Update one of the albums that you created earlier, or create one with the POST endpoint.

Be sure the Content-Type header is set to application/json so that Laravel knows the request body contains JSON and parses it correctly when you use $request->input(). This should happen automatically in Insomnia when you select Body > JSON.

https://laravel.com/docs/11.x/requests#retrieving-json-input-values

Sample request body: {"Title":"Test album (updated)","ArtistId":2}